sorting/adding features

+++ OSHW2/sort-files.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 #sort-files
-cd files
-exts=$(ls -l |awk -F'.' '{print $2}'|sort|uniq|xargs)
-for ext in $exts; do
-	mkdir $ext
-	for i in {1..10}
-	do
-		mv f$i.$ext $ext
+if [ -d "files"  ]; then #if files Dir exists
+	exts=$(ls -l |awk -F'.' '{print $2}'|sort|uniq|xargs) #get extenssions
+	for ext in $exts; do #loop throught the extenssions
+		mkdir $ext #mkdir for each exten
+		for i in {1..10}; do #loop for each file
+			mv f$i.$ext $ext # move it to the exten directory
+		done
 	done
-done
-ls -R #list dir and sub-contents
+	ls -R #list dir and sub-contents
+fi
